% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logica.R
\name{p_HDMT}
\alias{p_HDMT}
\title{Composite Null Hypothesis Testing}
\usage{
p_HDMT(p1, p2)
}
\arguments{
\item{sumstat_1}{Summary statistics for population 1.}

\item{sumstat_2}{Summary statistics for population 2.}

\item{R1}{LD matrix for population 1.}

\item{R2}{LD matrix for population 2.}

\item{n1}{Sample size for population 1.}

\item{n2}{Sample size for population 2.}

\item{...}{Additional parameters.}
}
\value{
A data frame with results.
}
\description{
This function runs the Logica pipeline using summary statistics and LD matrices.
}
