% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logica.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data for Logica}
\usage{
preprocess_data(z1, z2, R1, R2)
}
\arguments{
\item{z1}{Data frame or matrix for population 1.}

\item{z2}{Data frame or matrix for population 2.}

\item{R1}{LD matrix for population 1.}

\item{R2}{LD matrix for population 2.}
}
\value{
A list containing updated data and LD matrices.
}
\description{
This function preprocesses data by removing SNPs based on diagnostic criteria and aligning SNPs across populations.
}
