% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logica.R
\name{run_Logica_initial}
\alias{run_Logica_initial}
\title{Run Logica Analysis with different initials}
\usage{
run_Logica_initial(
  sumstat_1,
  sumstat_2,
  R1,
  R2,
  n1,
  n2,
  z1_intercept = 1,
  z2_intercept = 1,
  h2_1_initial,
  h2_2_initial,
  rho_initial,
  fix_intercept = TRUE
)
}
\arguments{
\item{sumstat_1}{Summary statistics for population 1.}

\item{sumstat_2}{Summary statistics for population 2.}

\item{R1}{LD matrix for population 1.}

\item{R2}{LD matrix for population 2.}

\item{n1}{Sample size for population 1.}

\item{n2}{Sample size for population 2.}

\item{...}{Additional parameters.}
}
\value{
A data frame with results.
}
\description{
This function runs the Logica pipeline using summary statistics and LD matrices.
}
